批量添加窗口中提炼出FragPagerAdapter
